- some more cout statements about loading things and such  - some crap to detect/prevent index overflows that i don't really like.. (should just convert to 64 bits IF it's a problem)
git-svn-id: file:///home/scott/regretitron_publish/regretitron_filtered3_svn/trunk@76 56bb13ec-af9c-4a7e-874f-62681e7d7c12

+++ PokerPlayerMFC/strategy.cpp
@@ -130,7 +130,7 @@ void Strategy::getprobs(int gr, int actioni, int numa, const vector<CardMask> &c
 	
 	unsigned char charprobs[MAX_ACTIONS];
 	//seekg ( offset + combinedindex * sizeofeachdata )
-	strategyfile.seekg( dataoffset[gr][numa-2] + COMBINE(cardsi, actioni, actionmax[gr][numa-2]) * numa );
+	strategyfile.seekg( dataoffset[gr][numa-2] + COMBINE(cardsi, (int64)actioni, actionmax[gr][numa-2]) * (int64)numa );
 	strategyfile.read((char*)charprobs, numa);
 	if(strategyfile.gcount()!=numa || strategyfile.eof())
 		REPORT("strategy file has failed us.");
